+++ lib/core/models/editor_configs/filter_editor_configs.dart
@@ -1,5 +1,4 @@
 import '/features/filter_editor/utils/filter_generator/filter_model.dart';
-
 import '../custom_widgets/filter_editor_widgets.dart';
 import '../icons/filter_editor_icons.dart';
 import '../styles/filter_editor_style.dart';
@@ -33,6 +32,7 @@ class FilterEditorConfigs implements BaseSubEditorConfigs {
   const FilterEditorConfigs({
     this.enableGesturePop = true,
     this.showLayers = true,
+    this.enableMultiSelection = true,
     this.filterList,
     this.safeArea = const EditorSafeArea(),
     this.fadeInUpDuration = const Duration(milliseconds: 220),
@@ -48,6 +48,11 @@ class FilterEditorConfigs implements BaseSubEditorConfigs {
 
   /// Show also layers in the editor.
   final bool showLayers;
+
+  /// When enabled, each time the editor is opened, a new filter will be
+  /// overlaid on top of the existing one. When disabled, selecting a filter
+  /// will replace the current one instead of stacking them.
+  final bool enableMultiSelection;
 
   /// A list of color filter generators to apply to an image.
   final List<FilterModel>? filterList;
@@ -87,6 +92,7 @@ class FilterEditorConfigs implements BaseSubEditorConfigs {
   FilterEditorConfigs copyWith({
     bool? enableGesturePop,
     bool? showLayers,
+    bool? enableMultiSelection,
     List<FilterModel>? filterList,
     Duration? fadeInUpDuration,
     Duration? fadeInUpStaggerDelayDuration,
@@ -97,13 +103,14 @@ class FilterEditorConfigs implements BaseSubEditorConfigs {
   }) {
     return FilterEditorConfigs(
       enableGesturePop: enableGesturePop ?? this.enableGesturePop,
-      safeArea: safeArea ?? this.safeArea,
       showLayers: showLayers ?? this.showLayers,
+      enableMultiSelection: enableMultiSelection ?? this.enableMultiSelection,
       filterList: filterList ?? this.filterList,
       fadeInUpDuration: fadeInUpDuration ?? this.fadeInUpDuration,
-      style: style ?? this.style,
       fadeInUpStaggerDelayDuration:
           fadeInUpStaggerDelayDuration ?? this.fadeInUpStaggerDelayDuration,
+      safeArea: safeArea ?? this.safeArea,
+      style: style ?? this.style,
       icons: icons ?? this.icons,
       widgets: widgets ?? this.widgets,
     );