+++ lib/core/models/editor_configs/text_editor_configs.dart
@@ -34,38 +34,39 @@ class TextEditorConfigs
   ///
   /// By default, the text editor is enabled, and most text formatting options
   /// are enabled. The initial font size is set to 24.0.
-  const TextEditorConfigs({
-    this.layerFractionalOffset = const Offset(-0.5, -0.5),
-    this.enableGesturePop = true,
-    this.enableSuggestions = true,
-    this.enableEdit = true,
-    this.enableAutocorrect = true,
-    this.showSelectFontStyleBottomBar = false,
-    this.showTextAlignButton = true,
-    this.showFontScaleButton = true,
-    this.showBackgroundModeButton = true,
-    this.enableMainEditorZoomFactor = false,
-    this.enableAutoOverflow = true,
-    this.enableImageBoundaryTextWrap = false,
-    this.enableTapOutsideToSave = true,
-    this.initFontSize = 24.0,
-    this.initialPrimaryColor = const Color(0xFF000000),
-    this.initialSecondaryColor,
-    this.initialTextAlign = TextAlign.center,
-    this.inputTextFieldAlign = Alignment.center,
-    this.initFontScale = 1.0,
-    this.maxFontScale = 3.0,
-    this.minFontScale = 0.3,
-    this.minScale = double.negativeInfinity,
-    this.maxScale = double.infinity,
-    this.customTextStyles,
-    this.defaultTextStyle = const TextStyle(),
-    this.initialBackgroundColorMode = LayerBackgroundMode.backgroundAndColor,
-    this.safeArea = const EditorSafeArea(),
-    this.style = const TextEditorStyle(),
-    this.icons = const TextEditorIcons(),
-    this.widgets = const TextEditorWidgets(),
-  })  : assert(initFontSize > 0, 'initFontSize must be positive'),
+  const TextEditorConfigs(
+      {this.layerFractionalOffset = const Offset(-0.5, -0.5),
+      this.enableGesturePop = true,
+      this.enableSuggestions = true,
+      this.enableEdit = true,
+      this.enableAutocorrect = true,
+      this.showSelectFontStyleBottomBar = false,
+      this.showTextAlignButton = true,
+      this.showFontScaleButton = true,
+      this.showBackgroundModeButton = true,
+      this.enableMainEditorZoomFactor = false,
+      this.enableTapOutsideToSave = true,
+      this.enableAutoOverflow = true,
+      this.enableAutoWrapOnLayer = true,
+      this.initFontSize = 24.0,
+      this.initialPrimaryColor = const Color(0xFF000000),
+      this.initialSecondaryColor,
+      this.initialTextAlign = TextAlign.center,
+      this.inputTextFieldAlign = Alignment.center,
+      this.initFontScale = 1.0,
+      this.maxFontScale = 3.0,
+      this.minFontScale = 0.3,
+      this.minScale = double.negativeInfinity,
+      this.maxScale = double.infinity,
+      this.customTextStyles,
+      this.defaultTextStyle = const TextStyle(),
+      this.initialBackgroundColorMode = LayerBackgroundMode.backgroundAndColor,
+      this.safeArea = const EditorSafeArea(),
+      this.style = const TextEditorStyle(),
+      this.icons = const TextEditorIcons(),
+      this.widgets = const TextEditorWidgets(),
+      this.enableImageBoundaryTextWrap = false})
+      : assert(initFontSize > 0, 'initFontSize must be positive'),
         assert(maxScale >= minScale,
             'maxScale must be greater than or equal to minScale');
 
@@ -152,6 +153,17 @@ class TextEditorConfigs
   /// (e.g., the screen width).
   final bool enableAutoOverflow;
 
+  /// Whether the text should automatically wrap when it reaches the end of
+  /// the screen on the final image.
+  ///
+  /// If set to `true`, the text will wrap to the next line instead of
+  /// overflowing, ensuring it stays within the visible area
+  /// (e.g., the screen width).
+  ///
+  /// If set to `false`, the text will only wrap if the user deliberately
+  /// entered a new line while editing.
+  final bool enableAutoWrapOnLayer;
+
   /// The minimum scale factor from the layer.
   final double minScale;
 
@@ -202,6 +214,7 @@ class TextEditorConfigs
     bool? enableMainEditorZoomFactor,
     bool? enableTapOutsideToSave,
     bool? enableAutoOverflow,
+    bool? enableAutoWrapOnLayer,
     Color? initialPrimaryColor,
     Color? initialSecondaryColor,
     double? initFontSize,
@@ -239,6 +252,8 @@ class TextEditorConfigs
       enableTapOutsideToSave:
           enableTapOutsideToSave ?? this.enableTapOutsideToSave,
       enableAutoOverflow: enableAutoOverflow ?? this.enableAutoOverflow,
+      enableAutoWrapOnLayer:
+          enableAutoWrapOnLayer ?? this.enableAutoWrapOnLayer,
       initialPrimaryColor: initialPrimaryColor ?? this.initialPrimaryColor,
       initialSecondaryColor:
           initialSecondaryColor ?? this.initialSecondaryColor,