+++ lib/features/paint_editor/models/path_builder/path_builder_base.dart
@@ -8,6 +8,7 @@ import 'path_builder_circle.dart';
 import 'path_builder_dash_dot_line.dart';
 import 'path_builder_dash_line.dart';
 import 'path_builder_freestyle.dart';
+import 'path_builder_hexagon.dart';
 import 'path_builder_line.dart';
 import 'path_builder_polygon.dart';
 import 'path_builder_rectangular.dart';
@@ -67,6 +68,14 @@ abstract class PathBuilderBase {
           item: item,
           scale: scale,
         );
+
+      case PaintMode.hexagon:
+        return PathBuilderHexagon(
+          paintEditorConfigs: paintEditorConfigs,
+          item: item,
+          scale: scale,
+        );
+
       case PaintMode.polygon:
         return PathBuilderPolygon(
           paintEditorConfigs: paintEditorConfigs,