+++ lib/plugins/image/src/exif/ifd_directory.dart
@@ -170,18 +170,14 @@ class IfdDirectory {
             case IfdValueType.single:
               if (value is List<double>) {
                 data[tag] = IfdValueSingle.list(Float32List.fromList(value));
-              } else if (value is double) {
-                data[tag] = IfdValueSingle(value);
-              } else if (value is int) {
+              } else if (value is num) {
                 data[tag] = IfdValueSingle(value.toDouble());
               }
               break;
             case IfdValueType.double:
               if (value is List<double>) {
                 data[tag] = IfdValueDouble.list(Float64List.fromList(value));
-              } else if (value is double) {
-                data[tag] = IfdValueDouble(value);
-              } else if (value is int) {
+              } else if (value is num) {
                 data[tag] = IfdValueDouble(value.toDouble());
               }
               break;